requirements.txt
python-picamera
